[#123] Move session into lib
