+++ lib/sessions/sessions_suite_test.go
@@ -1,0 +1,19 @@
+package sessions_test
+
+import (
+	"testing"
+
+	"go-google-scraper-challenge/test"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestWebControllers(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Sessions Suite")
+}
+
+var _ = BeforeSuite(func() {
+	test.SetupTestEnvironment()
+})
