+++ test/controller.go
@@ -12,7 +12,7 @@ import (
 
 	"go-google-scraper-challenge/helpers/log"
 	"go-google-scraper-challenge/lib/models"
-	"go-google-scraper-challenge/sessions"
+	"go-google-scraper-challenge/lib/sessions"
 
 	"github.com/gin-gonic/gin"
 	"github.com/gorilla/securecookie"
